fix time

+++ image_filter/mpi/blurmain.cc
@@ -313,7 +313,7 @@ int main(int argc, char** argv)
   clock_gettime(CLOCK_REALTIME, &etime);
 
   if(myid == 0){
-    const int delta_time = ((etime.tv_sec - stime.tv_sec) 
+    const double delta_time = ((etime.tv_sec - stime.tv_sec) 
 			    + 1e-9*(etime.tv_nsec - stime.tv_nsec));
     cout << "Filtering took: " <<  delta_time << "secs" << endl;
